move http objects and add global usings
