+++ Freelance_WbSellerFeedbackPareser/Http/IRequestSender.cs
@@ -1,0 +1,7 @@
+﻿namespace Freelance_WbSellerFeedbackPareser.Http
+{
+    internal interface IRequestSender
+    {
+        public Task<string> SendRequestAsync(HttpMethod method, string url);
+    }
+}
