+++ Freelance_WbSellerFeedbackPareser/Http/WbSellerHttpSender.cs
@@ -1,12 +1,4 @@
-﻿using Freelance_WbSellerFeedbackPareser.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Newtonsoft.Json.Linq;
-
-namespace Freelance_WbSellerFeedbackPareser
+﻿namespace Freelance_WbSellerFeedbackPareser.Http
 {
     internal class WbSellerHttpSender : WbSellerHttpClient, IRequestSender
     {
