+++ Freelance_WbSellerFeedbackPareser/WbAppealReader.cs
@@ -1,10 +1,4 @@
-﻿using Freelance_WbSellerFeedbackPareser.Models;
-using Newtonsoft.Json.Linq;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using Newtonsoft.Json.Linq;
 
 namespace Freelance_WbSellerFeedbackPareser
 {
