+++ Freelance_WbSellerFeedbackPareser/WbExcelAppealBuilder.cs
@@ -1,11 +1,4 @@
-﻿using Freelance_WbSellerFeedbackPareser.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Freelance_WbSellerFeedbackPareser
+﻿namespace Freelance_WbSellerFeedbackPareser
 {
     internal class WbExcelAppealBuilder
     {
